refactor(rules): extract rules code to dedicated module and reduce complexity
- Extract 480 lines of rules support from client.py to core/claude_rules.py
- Refactor _parse_rule_frontmatter() into focused helper functions:
  - _parse_inline_array()
  - _parse_paths_section()
  - _parse_skill_entry()
  - _parse_skills_section()
- Extract VALID_WHEN_VALUES and DEFAULT_WHEN_VALUE constants
- Fix CMT-001/002: Return empty string for invalid frontmatter
- Fix CMT-003: Implement skill-level path filtering
- Rename test_client.py to test_claude_rules.py

Addresses review findings:
- [HIGH] Cyclomatic complexity in _parse_rule_frontmatter()
- [MEDIUM] Module extraction to separate file
- [MEDIUM] Test assertion mismatches
- [MEDIUM] Skill-level paths not applied
