+++ apps/backend/core/client.py
@@ -13,16 +13,19 @@ single source of truth for phase-aware tool and MCP server configuration.
 """
 
 import copy
-import fnmatch
 import json
 import logging
 import os
-import re
 import threading
 import time
 from pathlib import Path
 from typing import Any
 
+from core.claude_rules import (
+    get_files_from_implementation_plan,
+    load_claude_rules,
+    should_use_claude_rules,
+)
 from core.platform import (
     is_windows,
     validate_cli_path,
@@ -444,485 +447,8 @@ def load_claude_md(project_dir: Path) -> str | None:
     return None
 
 
-# =============================================================================
-# Claude Code Rules Support (.claude/rules/)
-# =============================================================================
-# Supports Claude Code's path-based rules convention where rules in .claude/rules/
-# are automatically loaded based on which files are being modified.
-#
-# Rule files use YAML frontmatter with 'paths' array containing glob patterns:
-#
-#   ---
-#   paths:
-#     - src/app/api/**/*.ts
-#     - src/components/**/*.tsx
-#   ---
-#
-#   # Rule Content Here
-#   ...
-
-
-def should_use_claude_rules() -> bool:
-    """
-    Check if .claude/rules/ should be loaded based on file paths.
-
-    When USE_CLAUDE_MD is enabled, rules are also enabled by default.
-    Can be explicitly disabled with USE_CLAUDE_RULES=false.
-    """
-    explicit_setting = os.environ.get("USE_CLAUDE_RULES", "").lower()
-    if explicit_setting == "false":
-        return False
-    if explicit_setting == "true":
-        return True
-    # Default: enabled if USE_CLAUDE_MD is enabled
-    return should_use_claude_md()
-
-
-def _parse_rule_frontmatter(content: str) -> tuple[list[str], list[dict], str]:
-    """
-    Parse YAML frontmatter from a rule file to extract path patterns and required skills.
-
-    Args:
-        content: Full content of the rule file
-
-    Returns:
-        Tuple of (path_patterns, required_skills, rule_content_without_frontmatter)
-
-        required_skills is a list of dicts with format:
-        {"skill": "/review", "when": "end_of_coding", "paths": ["src/**"]}
-
-        Supported 'when' values:
-        - "planning": Planner should include in implementation_plan.json
-        - "per_subtask": Coder runs on each matching subtask (default)
-        - "end_of_coding": Coder runs once after ALL subtasks complete
-        - "qa_phase": QA agent runs during review
-
-    Supported frontmatter formats:
-        # Simple (backwards compatible) - defaults to when: per_subtask
-        require_skills:
-          - /review
-          - /audit
-
-        # Structured with timing control
-        require_skills:
-          - skill: /review
-            when: end_of_coding
-          - skill: /security-audit
-            when: per_subtask
-            paths: [src/app/api/**]
-    """
-    if not content.startswith("---"):
-        return [], [], content
-
-    # Find the closing ---
-    lines = content.split("\n")
-    end_idx = -1
-    for i, line in enumerate(lines[1:], 1):
-        if line.strip() == "---":
-            end_idx = i
-            break
-
-    if end_idx == -1:
-        return [], [], content
-
-    # Parse the frontmatter
-    frontmatter_lines = lines[1:end_idx]
-    paths = []
-    skills: list[dict] = []
-
-    in_paths = False
-    in_skills = False
-    in_skill_paths = False  # Track when parsing nested paths: list under a skill
-    current_skill: dict | None = None  # Track structured skill entry
-    skill_base_indent: int = 0  # Track base indent of current skill list item
-
-    for line in frontmatter_lines:
-        stripped = line.strip()
-        # Calculate indentation for detecting nested properties
-        indent = len(line) - len(line.lstrip())
-
-        # Parse paths: array
-        if stripped.startswith("paths:") and not in_skills:
-            in_paths = True
-            in_skills = False
-            current_skill = None
-            # Check for inline array: paths: [a, b, c]
-            if "[" in stripped:
-                match = re.search(r"\[(.*)\]", stripped)
-                if match:
-                    paths = [p.strip().strip("'\"") for p in match.group(1).split(",")]
-                in_paths = False
-            continue
-        # Parse require_skills: array
-        if stripped.startswith("require_skills:"):
-            in_skills = True
-            in_paths = False
-            current_skill = None
-            # Check for inline array: require_skills: [/skill1, /skill2]
-            if "[" in stripped:
-                match = re.search(r"\[(.*)\]", stripped)
-                if match:
-                    # Simple inline format - convert to dicts with default 'when'
-                    for s in match.group(1).split(","):
-                        skill_name = s.strip().strip("'\"")
-                        if skill_name:
-                            skills.append({"skill": skill_name, "when": "per_subtask"})
-                in_skills = False
-            continue
-        if in_paths:
-            if stripped.startswith("- "):
-                paths.append(stripped[2:].strip().strip("'\""))
-                continue
-            elif stripped and not stripped.startswith("#"):
-                # End of paths array - fall through to check for other keys
-                in_paths = False
-        if in_skills:
-            # Collect nested paths list items for current skill
-            if (
-                current_skill is not None
-                and in_skill_paths
-                and stripped.startswith("- ")
-                and indent > skill_base_indent
-            ):
-                current_skill.setdefault("paths", []).append(
-                    stripped[2:].strip().strip("'\"")
-                )
-                continue
-
-            # Check for new list item
-            if stripped.startswith("- "):
-                # Save previous skill entry if exists
-                if current_skill and current_skill.get("skill"):
-                    skills.append(current_skill)
-                    current_skill = None
-                in_skill_paths = False
-
-                item_content = stripped[2:].strip()
-                # Track the indent of this list item for nested property detection
-                skill_base_indent = indent
-
-                # Check if it's structured format: "- skill: /review"
-                if item_content.startswith("skill:"):
-                    skill_name = item_content[6:].strip().strip("'\"")
-                    current_skill = {"skill": skill_name, "when": "per_subtask"}
-                else:
-                    # Simple format: "- /review" - convert to dict immediately
-                    skill_name = item_content.strip("'\"")
-                    if skill_name:
-                        skills.append({"skill": skill_name, "when": "per_subtask"})
-            # Check for nested properties of current skill (indented more than list item)
-            elif current_skill is not None and indent > skill_base_indent:
-                if stripped.startswith("when:"):
-                    when_value = stripped[5:].strip().strip("'\"")
-                    if when_value in (
-                        "planning",
-                        "per_subtask",
-                        "end_of_coding",
-                        "qa_phase",
-                    ):
-                        current_skill["when"] = when_value
-                elif stripped.startswith("paths:"):
-                    in_skill_paths = True
-                    current_skill.setdefault("paths", [])
-                    # Check for inline array: paths: [src/**, lib/**]
-                    if "[" in stripped:
-                        match = re.search(r"\[(.*)\]", stripped)
-                        if match:
-                            current_skill["paths"] = [
-                                p.strip().strip("'\"")
-                                for p in match.group(1).split(",")
-                            ]
-                        in_skill_paths = False
-            elif (
-                stripped
-                and not stripped.startswith("#")
-                and indent <= skill_base_indent
-            ):
-                # End of skills array (same or less indent, non-empty, non-comment line)
-                if current_skill and current_skill.get("skill"):
-                    skills.append(current_skill)
-                    current_skill = None
-                in_skills = False
-                in_skill_paths = False
-                # Reprocess this line - it might be a top-level paths: that follows require_skills:
-                if stripped.startswith("paths:") and not in_paths:
-                    in_paths = True
-                    # Check for inline array: paths: [a, b, c]
-                    if "[" in stripped:
-                        match = re.search(r"\[(.*)\]", stripped)
-                        if match:
-                            paths = [
-                                p.strip().strip("'\"")
-                                for p in match.group(1).split(",")
-                            ]
-                        in_paths = False
-
-    # Don't forget to save the last skill entry
-    if current_skill and current_skill.get("skill"):
-        skills.append(current_skill)
-
-    # Return content without frontmatter
-    rule_content = "\n".join(lines[end_idx + 1 :]).strip()
-    return paths, skills, rule_content
-
-
-def _match_glob_pattern(pattern: str, filepath: str) -> bool:
-    """
-    Match a glob pattern against a file path.
-
-    Supports:
-    - ** for any directory depth (including multiple ** in a pattern)
-    - * for any characters within a path segment
-    - Direct path matching
-
-    Args:
-        pattern: Glob pattern (e.g., "src/app/api/**/*.ts")
-        filepath: File path to check (e.g., "src/app/api/films/route.ts")
-
-    Returns:
-        True if the pattern matches the filepath
-    """
-    # Normalize paths
-    pattern = pattern.replace("\\", "/").strip("/")
-    filepath = filepath.replace("\\", "/").strip("/")
-
-    # Handle ** patterns by converting to regex
-    if "**" in pattern:
-        # Convert glob pattern to regex
-        # Escape regex special characters except * and ?
-        regex_pattern = ""
-        i = 0
-        while i < len(pattern):
-            if pattern[i : i + 2] == "**":
-                i += 2
-                # Check if ** is followed by /
-                if i < len(pattern) and pattern[i] == "/":
-                    # **/ matches zero or more directory segments (segment-aware)
-                    # Use (?:.*/)? to match zero or more path segments including slashes
-                    regex_pattern += "(?:.*/)?"
-                    i += 1  # Skip the /
-                else:
-                    # ** at end or before non-/ matches remaining path
-                    regex_pattern += ".*"
-            elif pattern[i] == "*":
-                # * matches any characters except /
-                regex_pattern += "[^/]*"
-                i += 1
-            elif pattern[i] == "?":
-                # ? matches any single character except /
-                regex_pattern += "[^/]"
-                i += 1
-            elif pattern[i] in ".^$+{}[]|()":
-                # Escape regex special characters
-                regex_pattern += "\\" + pattern[i]
-                i += 1
-            else:
-                regex_pattern += pattern[i]
-                i += 1
-
-        # Match the full path
-        regex_pattern = "^" + regex_pattern + "$"
-        return bool(re.match(regex_pattern, filepath))
-
-    # Simple glob matching for patterns without **
-    return fnmatch.fnmatch(filepath, pattern)
-
-
-def _discover_rules_directory(project_dir: Path) -> Path | None:
-    """
-    Find the .claude/rules/ directory if it exists.
-
-    Args:
-        project_dir: Root directory of the project
-
-    Returns:
-        Path to rules directory if found, None otherwise
-    """
-    rules_dir = project_dir / ".claude" / "rules"
-    if not rules_dir.exists() or not rules_dir.is_dir():
-        return None
-
-    # Security: Verify the resolved path is still under project_dir
-    # This prevents symlink attacks where .claude or rules points outside
-    try:
-        resolved_rules = rules_dir.resolve()
-        resolved_project = project_dir.resolve()
-        if not str(resolved_rules).startswith(str(resolved_project) + os.sep):
-            logger.warning(
-                f"Rules directory {rules_dir} resolves outside project: {resolved_rules}"
-            )
-            return None
-    except (OSError, ValueError) as e:
-        logger.warning(f"Failed to resolve rules directory path: {e}")
-        return None
-
-    return resolved_rules
-
-
-def _collect_all_rules(
-    rules_dir: Path,
-) -> list[tuple[Path, list[str], list[dict], str]]:
-    """
-    Recursively collect all rule files from the rules directory.
-
-    Args:
-        rules_dir: Path to .claude/rules/
-
-    Returns:
-        List of tuples: (rule_path, path_patterns, required_skills, rule_content)
-        where required_skills is a list of dicts with 'skill', 'when', and optional 'paths' keys
-    """
-    rules = []
-    rules_dir_resolved = rules_dir.resolve()
-
-    for rule_path in rules_dir.rglob("*.md"):
-        # Skip symlink files to prevent traversal outside project directory
-        if rule_path.is_symlink():
-            logger.warning(f"Skipping symlink rule file: {rule_path}")
-            continue
-
-        # Resolve the path and verify it stays under rules_dir
-        # (handles symlinked parent directories)
-        try:
-            rule_path_resolved = rule_path.resolve()
-            rule_path_resolved.relative_to(rules_dir_resolved)
-        except (OSError, RuntimeError, ValueError) as e:
-            logger.warning(
-                f"Skipping rule file that failed to resolve or escaped rules dir: {rule_path} ({e})"
-            )
-            continue
-
-        try:
-            content = rule_path_resolved.read_text(encoding="utf-8")
-            paths, skills, rule_content = _parse_rule_frontmatter(content)
-            if paths and rule_content:
-                rules.append((rule_path, paths, skills, rule_content))
-            elif rule_content and not paths:
-                logger.debug(
-                    f"Skipping rule {rule_path}: no paths defined in frontmatter"
-                )
-        except Exception as e:
-            logger.debug(f"Failed to read rule file {rule_path}: {e}")
-
-    return rules
-
-
-def load_claude_rules(
-    project_dir: Path,
-    files_to_check: list[str] | None = None,
-) -> tuple[str, list[str], list[dict]]:
-    """
-    Load Claude Code rules from .claude/rules/ that match the given files.
-
-    Rules are markdown files with YAML frontmatter containing path patterns
-    and optional required skills.
-
-    Only rules whose patterns match at least one file in files_to_check are loaded.
-
-    Args:
-        project_dir: Root directory of the project
-        files_to_check: List of file paths being modified/created.
-                       If None, loads ALL rules (useful for planning phases).
-
-    Returns:
-        Tuple of (combined_rules_content, list_of_matched_rule_names, list_of_required_skills)
-        where required_skills is a list of dicts with 'skill', 'when', and optional 'paths' keys
-    """
-    rules_dir = _discover_rules_directory(project_dir)
-    if not rules_dir:
-        return "", [], []
-
-    all_rules = _collect_all_rules(rules_dir)
-    if not all_rules:
-        return "", [], []
-
-    matched_rules = []
-    matched_names = []
-    # Track skills by (name, when, paths) to preserve different configurations
-    # of the same skill across phases or with different path filters
-    seen_skills: set[tuple[str, str, tuple[str, ...]]] = set()
-    all_required_skills: list[dict] = []
-
-    def _add_skill(skill: dict) -> None:
-        """Add skill if not already seen with same configuration."""
-        skill_name = skill.get("skill", "")
-        if not skill_name:
-            return
-        when = skill.get("when", "per_subtask")
-        paths_key = tuple(sorted(skill.get("paths", []) or []))
-        key = (skill_name, when, paths_key)
-        if key not in seen_skills:
-            seen_skills.add(key)
-            all_required_skills.append(skill)
-
-    for rule_path, patterns, skills, content in all_rules:
-        # Get relative name for display
-        rel_name = str(rule_path.relative_to(rules_dir))
-
-        # If no files specified, load all rules
-        if files_to_check is None:
-            matched_rules.append((rel_name, content))
-            matched_names.append(rel_name)
-            for skill in skills:
-                _add_skill(skill)
-            continue
-
-        # Check if any pattern matches any file
-        for pattern in patterns:
-            for filepath in files_to_check:
-                if _match_glob_pattern(pattern, filepath):
-                    matched_rules.append((rel_name, content))
-                    matched_names.append(rel_name)
-                    for skill in skills:
-                        _add_skill(skill)
-                    break
-            else:
-                continue
-            break
-
-    if not matched_rules:
-        return "", [], []
-
-    # Combine matched rules into a single string
-    combined = []
-    for name, content in matched_rules:
-        combined.append(f"## Rule: {name}\n\n{content}")
-
-    return "\n\n---\n\n".join(combined), matched_names, all_required_skills
-
-
-def _get_files_from_implementation_plan(spec_dir: Path) -> list[str]:
-    """
-    Extract all files_to_modify and files_to_create from implementation_plan.json.
-
-    Args:
-        spec_dir: Directory containing the spec
-
-    Returns:
-        List of all file paths mentioned in the plan
-    """
-    plan_path = spec_dir / "implementation_plan.json"
-    if not plan_path.exists():
-        return []
-
-    try:
-        plan = json.loads(plan_path.read_text(encoding="utf-8"))
-        files = set()
-
-        # Collect from all phases and subtasks
-        for phase in plan.get("phases", []):
-            for subtask in phase.get("subtasks", []):
-                files.update(subtask.get("files_to_modify", []))
-                files.update(subtask.get("files_to_create", []))
-
-        # Also check top-level if present
-        files.update(plan.get("files_to_modify", []))
-        files.update(plan.get("files_to_create", []))
-
-        return list(files)
-    except Exception as e:
-        logger.debug(f"Failed to read implementation plan: {e}")
-        return []
+# Note: Claude Code Rules support is now in core/claude_rules.py
+# Exported functions: should_use_claude_rules, load_claude_rules, get_files_from_implementation_plan
 
 
 def create_client(
@@ -1281,7 +807,7 @@ def create_client(
     # Rules are matched based on files being modified in the implementation plan
     required_skills: list[dict] = []
     if should_use_claude_rules():
-        files_in_plan = _get_files_from_implementation_plan(spec_dir)
+        files_in_plan = get_files_from_implementation_plan(spec_dir)
         if files_in_plan:
             rules_content, matched_rules, required_skills = load_claude_rules(
                 project_dir, files_in_plan
